+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### features
 * allow to return a promise from greetings function [#531](https://github.com/jcubic/jquery.terminal/issues/531)
 * call onClear before clear and allow to cancel [#527](https://github.com/jcubic/jquery.terminal/issues/527)
+* new render method and renderHandler option [#526](https://github.com/jcubic/jquery.terminal/issues/526)
 ### Bugfix
 * fix links from formatters in cmd
 * fix images from formatters in cmd [#519](https://github.com/jcubic/jquery.terminal/issues/519)